[#37] Refactored some js folder and file names, locations
