+++ src/common/tableCell.module.scss
