+++ src/components/players/playerDetails/stats/stats.js
@@ -1,0 +1,1 @@
+import React from 'react';
